@@ -2,43 +2,41 @@
   "name": "example-project",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "jest",
     "start": "node src/server/index.js",
-    "build-prod": "SET NODE_OPTIONS=--openssl-legacy-provider && webpack --config webpack.prod.js",
-    "build-dev": "SET NODE_OPTIONS=--openssl-legacy-provider && webpack-dev-server  --config webpack.dev.js --open"
+    "build-prod": "cross-env NODE_OPTIONS=--openssl-legacy-provider webpack --config webpack.prod.cjs",
+    "build-dev": "cross-env NODE_OPTIONS=--openssl-legacy-provider webpack serve --config webpack.dev.cjs --open"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "description": "",
   "dependencies": {
-    "dotenv": "^8.2.0",
-    "express": "^4.17.1",
-    "jest-fetch-mock": "^3.0.3",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.5"
+    "dotenv": "^17.2.3",
+    "express": "^5.1.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.13.15",
-    "@babel/plugin-transform-modules-commonjs": "^7.13.8",
-    "@babel/preset-env": "^7.13.15",
-    "babel-loader": "^8.2.2",
-    "body-parser": "^1.19.0",
-    "clean-webpack-plugin": "^3.0.0",
+    "@babel/core": "^7.28.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.27.1",
+    "@babel/preset-env": "^7.28.5",
+    "babel-loader": "^10.0.0",
     "cors": "^2.8.5",
-    "css-loader": "^5.2.1",
-    "html-webpack-plugin": "^4.5.2",
-    "jest": "^26.6.3",
-    "mini-css-extract-plugin": "^0.12.0",
-    "node-fetch": "^2.6.1",
-    "node-sass": "^5.0.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.4",
-    "sass": "^1.32.8",
-    "sass-loader": "^10.1.1",
-    "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^4.2.3",
-    "webpack-dev-server": "^3.11.2",
-    "workbox-webpack-plugin": "^6.6.0"
+    "cross-env": "^10.1.0",
+    "css-loader": "^7.1.2",
+    "css-minimizer-webpack-plugin": "^7.0.2",
+    "html-webpack-plugin": "^5.6.4",
+    "jest": "^30.2.0",
+    "jest-fetch-mock": "^3.0.3",
+    "mini-css-extract-plugin": "^2.9.4",
+    "sass": "^1.93.3",
+    "sass-loader": "^16.0.6",
+    "style-loader": "^4.0.0",
+    "terser-webpack-plugin": "^5.3.14",
+    "webpack": "^5.102.1",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.2",
+    "workbox-webpack-plugin": "^7.3.0"
   }
 }
