@@ -1,61 +1,37 @@
-const dotenv = require('dotenv');
+import dotenv from 'dotenv';
+import path from 'path';
+import { fileURLToPath } from 'url';
+import express from 'express';
+import cors from 'cors';
+
+// Initialize dotenv to load environment variables.
 dotenv.config();
 
-const path = require('path');
-const express = require('express');
-const bodyParser = require('body-parser');
-const cors = require('cors');
-const fetch = require('node-fetch');
-const FormData = require('form-data');
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 const app = express();
 
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(bodyParser.json());
+app.use(express.json());
+app.use(express.urlencoded({ extended: false }));
+
 app.use(cors());
-app.use(express.static('dist'));
+app.use(express.static(path.join(__dirname, '../../dist')));
 
 app.get('/', function (req, res) {
-    res.sendFile('dist/index.html')
+    res.sendFile(path.resolve('dist/index.html'));
 });
 
-app.listen(8081, function () {
+
+const PORT = 8081;
+app.listen(PORT, function () {
+    console.log(`Server listening on port ${PORT}!`);
 });
 
 app.post('/analyzeText', analyzeText);
 
-async function analyzeText(req, res) {
-    const formData = new FormData();
-
-    formData.append("key", process.env.API_KEY);
-    formData.append("txt", req.body.text);
-    formData.append("lang", "en");
-
-    const requestOptions = {
-        method: 'POST',
-        body: formData,
-        redirect: 'follow'
-    };
-
-    try {
-        const response = await fetch("https://api.meaningcloud.com/sentiment-2.1", requestOptions);
-        const body = await response.json();
-
-        const responseData = {
-            subjectivity: body.subjectivity,
-            polarity: parsePolarity(body.score_tag),
-            text: body.sentence_list.length ? body.sentence_list[Math.floor(Math.random() * body.sentence_list.length)].text : req.body.text
-        }
-
-        res.send(responseData)
-
-    } catch (error) {
-        console.log('error', error);
-    }
-}
-
 function parsePolarity(sentiment) {
-    let sentimentMap = {
+    const sentimentMap = {
         "P+": "strong positive",
         "P": "positive",
         "NEU": "neutral",
@@ -63,8 +39,42 @@ function parsePolarity(sentiment) {
         "N+": "strong negative",
         "NONE": "without polarity"
     };
-
     return sentimentMap[sentiment];
+}
+
+async function analyzeText(req, res) {
+    // FormData is available globally in recent Node versions.
+    const formData = new FormData();
+    formData.append("key", process.env.API_KEY);
+    formData.append("txt", req.body.text);
+    formData.append("lang", "en");
+
+    const requestOptions = {
+        method: 'POST',
+        body: formData,
+    };
+
+    try {
+        const response = await fetch("https://api.meaningcloud.com/sentiment-2.1", requestOptions);
+
+        if (!response.ok) {
+            throw new Error(`API call failed with status: ${response.status}`);
+        }
+
+        const body = await response.json();
+
+        const responseData = {
+            subjectivity: body.subjectivity,
+            polarity: parsePolarity(body.score_tag),
+            text: body.sentence_list.length ? body.sentence_list[Math.floor(Math.random() * body.sentence_list.length)].text : req.body.text
+        };
+
+        res.send(responseData);
+
+    } catch (error) {
+        console.log('Error:', error.message);
+        res.status(500).send({ error: 'An error occurred while analyzing the text.' });
+    }
 }
 
 export { analyzeText, parsePolarity }
