@@ -2,7 +2,7 @@ const path = require('path')
 const webpack = require('webpack')
 const HtmlWebPackPlugin = require("html-webpack-plugin")
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const CssMinimizerPlugin = require('css-minimizer-webpack-plugin')
 const TerserPlugin = require('terser-webpack-plugin');
 const WorkboxPlugin = require('workbox-webpack-plugin');
 
@@ -11,24 +11,30 @@ module.exports = {
     mode: 'production',
     devtool: 'source-map',
     stats: 'verbose',
-    optimization: {
-        minimizer: [new TerserPlugin({}), new OptimizeCSSAssetsPlugin({})],
-    },
     output: {
-        libraryTarget: 'var',
-        library: 'Client'
+        library: {
+            name: 'Client',
+            type: 'var',
+        },
+        path: path.resolve(__dirname, 'dist'),
+        filename: '[name].[contenthash].js',
+        clean: true,
+    },
+    optimization: {
+        minimize: true,
+        minimizer: [new TerserPlugin({}), new CssMinimizerPlugin({}),({})],
     },
     module: {
         rules: [
             {
-                test: '/\.js$/',
+                test: /\.js$/,
                 exclude: /node_modules/,
                 loader: "babel-loader"
             },
             {
                 test: /\.scss$/,
                 use: [MiniCssExtractPlugin.loader, "css-loader", "sass-loader"],
-                },
+            },
         ]
     },
     plugins: [
@@ -37,7 +43,7 @@ module.exports = {
             filename: "./index.html",
         }),
         new MiniCssExtractPlugin({
-            filename: "[name].css"
+            filename: "[name].[contenthash].css"
         }),
         new WorkboxPlugin.GenerateSW()
     ]
